switched from rewrite to hack
got looking at it and not too much needs changed, so I'll just hijack
what I need instead of needlessly rewriting the whole game.
In this commit, I updated how player graphics are dma'd to vram, how the
controllers are polled, how player poses are handled, and how players
are drawn to the screen.